Use built-in PEM certificate import on .NET 6 onwards
This allows to drop the BouncyCastle.Cryptography dependency (almost 7 MB) when targeting net6.0 and net8.0.

Additionnally, the `Microsoft.Bcl.AsyncInterfaces` and `System.Text.Json` dependencies are also dropped when targeting .NET 6 and later.
