+++ src/Testcontainers/Builders/MTlsEndpointAuthenticationProvider.cs
@@ -1,25 +1,16 @@
 namespace DotNet.Testcontainers.Builders
 {
-  using System;
   using System.IO;
   using System.Linq;
   using System.Security.Cryptography.X509Certificates;
   using Docker.DotNet.X509;
   using DotNet.Testcontainers.Configurations;
   using JetBrains.Annotations;
-  using Org.BouncyCastle.Crypto;
-  using Org.BouncyCastle.Crypto.Parameters;
-  using Org.BouncyCastle.OpenSsl;
-  using Org.BouncyCastle.Pkcs;
-  using Org.BouncyCastle.Security;
-  using Org.BouncyCastle.X509;
 
   /// <inheritdoc cref="IDockerRegistryAuthenticationProvider" />
   [PublicAPI]
   internal sealed class MTlsEndpointAuthenticationProvider : TlsEndpointAuthenticationProvider
   {
-    private static readonly X509CertificateParser CertificateParser = new X509CertificateParser();
-
     /// <summary>
     /// Initializes a new instance of the <see cref="MTlsEndpointAuthenticationProvider" /> class.
     /// </summary>
@@ -57,57 +48,18 @@ namespace DotNet.Testcontainers.Builders
     {
       var clientCertificateFilePath = Path.Combine(CertificatesDirectoryPath, ClientCertificateFileName);
       var clientCertificateKeyFilePath = Path.Combine(CertificatesDirectoryPath, ClientCertificateKeyFileName);
-      return CreateFromPemFile(clientCertificateFilePath, clientCertificateKeyFilePath);
-    }
-
-    private static X509Certificate2 CreateFromPemFile(string certPemFilePath, string keyPemFilePath)
-    {
-      if (!File.Exists(certPemFilePath))
+#if NETSTANDARD
+      return Polyfill.X509Certificate2.CreateFromPemFile(clientCertificateFilePath, clientCertificateKeyFilePath);
+#else
+      var certificate = X509Certificate2.CreateFromPemFile(clientCertificateFilePath, clientCertificateKeyFilePath);
+      if (System.OperatingSystem.IsWindows())
       {
-        throw new FileNotFoundException(certPemFilePath);
+        // The certificate must be exported to PFX on Windows in order to avoid "No credentials are available in the security package"
+        // https://stackoverflow.com/questions/72096812/loading-x509certificate2-from-pem-file-results-in-no-credentials-are-available/72101855#72101855
+        return new X509Certificate2(certificate.Export(X509ContentType.Pfx));
       }
-
-      if (!File.Exists(keyPemFilePath))
-      {
-        throw new FileNotFoundException(keyPemFilePath);
-      }
-
-      using (var keyPairStream = new StreamReader(keyPemFilePath))
-      {
-        var store = new Pkcs12StoreBuilder().Build();
-
-        var certificate = CertificateParser.ReadCertificate(File.ReadAllBytes(certPemFilePath));
-
-        var password = Guid.NewGuid().ToString("D");
-
-        var keyObject = new PemReader(keyPairStream).ReadObject();
-
-        var certificateEntry = new X509CertificateEntry(certificate);
-
-        var keyParameter = ResolveKeyParameter(keyObject);
-
-        var keyEntry = new AsymmetricKeyEntry(keyParameter);
-        store.SetKeyEntry(certificate.SubjectDN + "_key", keyEntry, new[] { certificateEntry });
-
-        using (var certificateStream = new MemoryStream())
-        {
-          store.Save(certificateStream, password.ToCharArray(), new SecureRandom());
-          return new X509Certificate2(Pkcs12Utilities.ConvertToDefiniteLength(certificateStream.ToArray()), password);
-        }
-      }
-    }
-
-    private static AsymmetricKeyParameter ResolveKeyParameter(object keyObject)
-    {
-      switch (keyObject)
-      {
-        case AsymmetricCipherKeyPair ackp:
-          return ackp.Private;
-        case RsaPrivateCrtKeyParameters rpckp:
-          return rpckp;
-        default:
-          throw new ArgumentOutOfRangeException(nameof(keyObject), $"Unsupported asymmetric key entry encountered while trying to resolve key from input object '{keyObject.GetType()}'.");
-      }
+      return certificate;
+#endif
     }
   }
 }
