+++ tests/Testcontainers.Tests/Testcontainers.Tests.csproj
@@ -11,6 +11,7 @@
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk"/>
+    <PackageReference Include="BouncyCastle.Cryptography"/>
     <PackageReference Include="coverlet.collector"/>
     <PackageReference Include="xunit.runner.visualstudio"/>
     <PackageReference Include="xunit"/>
